Działa tak, jak dzialalo w ObjC, tyle że document based. Bez progress bara ale może inną metodą można

+++ ManagedContext/Event+CoreDataProperties.swift
@@ -15,7 +15,7 @@ import CoreData
 extension Event {
 
     @NSManaged var media: NSData?
-    @NSManaged var duration: Float
+    @NSManaged var duration: Double
     @NSManaged var mediaName: String?
     @NSManaged var title: String?
     @NSManaged var mediaType: String?
